♻️ refactor : 이벤트 배너 삭제

See merge request s10-final/S10P31C107!233

+++ backend/event/src/main/java/com/luckyseven/event/rollsheet/service/EventServiceImpl.java
@@ -185,7 +185,9 @@ public class EventServiceImpl implements EventService {
             }
         }
 
-        fileService.deleteBannerImageOnAmazonS3(eventId);
+        if (event.getBanner() != null) {
+            fileService.deleteBannerImageOnAmazonS3(eventId);
+        }
         eventRepository.delete(event);
     }
